requirements.txt
requests>=2.28
pdfplumber
python-docx
openpyxl
python-pptx
